Frontend/FrontSonInterface/requirements.txt
customtkinter==5.1.2
darkdetect==0.8.0
Pillow==11.2.1
